chore[react-devtools]: extract getDispatcherRef into a separate module
